reddit_airflow/requirements.txt
praw
apache-airflow[google]
langchain
langchain-pinecone
langchain-text-splitters
langchain-huggingface
pinecone
pinecone-notebooks